exam-ai/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.11.3
python-dotenv==1.1.0

langchain==0.3.24
langchain-core==0.3.56
langchain-google-genai==2.1.4
langgraph==0.3.31

chromadb==0.6.3

pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
joblib==1.4.2
matplotlib==3.10.1
seaborn==0.13.2
